Updated base algorithm
